fix(tls): move peer identity type to tls crate
