+++ filter/Cargo.toml
@@ -32,6 +32,7 @@ http = { workspace = true }
 metrics = { workspace = true }
 percent-encoding = { workspace = true }
 praxis-core = { workspace = true }
+praxis-tls = { workspace = true }
 rand = { workspace = true }
 regex = { workspace = true }
 serde = { workspace = true }
