+++ filter/src/lib.rs
@@ -34,12 +34,13 @@ pub use builtins::{
     normalize_rewritten_path,
 };
 pub use condition::{should_execute, should_execute_response, should_execute_response_ref};
-pub use context::{HttpFilterContext, PendingHeaderResult, Request, Response, TlsPeerIdentity, TrustedHeaderMutation};
+pub use context::{HttpFilterContext, PendingHeaderResult, Request, Response, TrustedHeaderMutation};
 pub use extensions::RequestExtensions;
 pub use factory::{FilterFactory, HttpFilterFactory, TcpFilterFactory, http_builtin, parse_filter_config, tcp_builtin};
 pub use filter::{Filter, FilterContext, FilterError, HttpFilter};
 pub use pipeline::{FilterPipeline, PipelineExtension};
 pub use praxis_core::config::{FailureMode, FilterEntry};
+pub use praxis_tls::TlsPeerIdentity;
 pub use registry::FilterRegistry;
 pub use results::FilterResultSet;
 pub use tcp_filter::{TcpFilter, TcpFilterContext};
